requirements.txt
tensorflow
streamlit
pillow
numpy
split-folders
kaggle
matplotlib